+++ src/main/java/com/imsweb/mph/mpgroups/Mp2010HematopoieticGroup.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 import com.imsweb.mph.MphConstants;
 import com.imsweb.mph.MphGroup;
@@ -446,7 +447,7 @@ public class Mp2010HematopoieticGroup extends MphGroup {
                 if (GroupUtility.differentCategory(i1.getHistology(), i2.getHistology(), MphConstants.PTLD, combined)) {
                     //Even if the dx date is known to be the same, we don't know if it is in one biopsy.
                     //If dx date is same, return a potential single primary! If M15 confirms the cases as single, we will return single.
-                    if (i1.getDateOfDiagnosisYear().equals(i2.getDateOfDiagnosisYear()) && i1.getDateOfDiagnosisMonth().equals(i2.getDateOfDiagnosisMonth()) && i1.getDateOfDiagnosisDay().equals(i2.getDateOfDiagnosisDay())) {
+                    if (i1.getDateOfDiagnosisYear().equals(i2.getDateOfDiagnosisYear()) && Objects.equals(i1.getDateOfDiagnosisMonth(), i2.getDateOfDiagnosisMonth()) && Objects.equals(i1.getDateOfDiagnosisDay(), i2.getDateOfDiagnosisDay())) {
                         result.setPotentialResult(MphUtils.MpResult.SINGLE_PRIMARY);
                         result.setMessageUnknownDiagnosisDate(this.getStep(), this.getGroupName());
                     }