+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Changes in version 1.33**
 
+- Cleaned up a code handling missing invalid date parts. (#128)
 - Fixed definition of simultaneous to be same day for M14 of Hematopoietic and Lymphoid 2010+ rules. (#129)
 - Return questionable for same row rules if the histologies being compared are the same, even if they are not in the table. (#130)